+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `SidebarCloseButton` component.
+- `sidebarCloseIconProps` prop on `filter-navigator.v2` and `filter-navigator.v3`.
 
 ## [3.52.0] - 2020-03-26
 ### Added